Return only lines ending with \n
linemux shouldn't return the last line of a file if it doesn't end with
a newline character. It could just be an incomplete line that will be
completed by subsequent writes to the file.

E.g. for the following writes

    echo -n "Hello" > tail.log
    echo -n "World\n" >> tail.log

linemux should only ever return a single line "Hello World" and never
two lines "Hello" and "World".